requirements.txt
# Hosting Jupyter Notebooks on GitHub Pages
# 
# Author:  Anshul Kharbanda
# Created: 10 - 12 - 2020

# Jupyter notebook environment
notebook ~= 6.1
numpy ~= 1.19
matplotlib ~= 3.3
scipy ~= 1.5
pandas ~= 1.1

# Site building
jinja2 ~= 2.11
markdown ~= 3.3
pyyaml ~= 5.3
click ~= 7.1